chore: settings windos iteration.
