+++ com.stansassets.facebook/Editor/FbSettingsTab.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c178494de8604666a63bdbb98b8a2713
+timeCreated: 1601727742
