+++ com.stansassets.facebook/Runtime/Config/FacebookAdditiveSettings.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d7cc297a24c74b4389bb20e61e77591d
+timeCreated: 1601744409
