+++ com.stansassets.facebook/package.json
@@ -3,7 +3,7 @@
   "displayName": "Stans Assets - Facebook",
   "version": "1.0.0",
   "unity": "2019.3",
-  "description": "Package Sample.",
+  "description": "Power up your Unity games with our Extended Facebook SDK. This library is a free open source product that contains extended access to Facebook API.",
   "dependencies": {
     "com.stansassets.foundation": "1.0.5",
     "com.stansassets.plugins-dev-kit" : "0.0.3-preview"
